Migrate GetPaymentNotificationMetadataRecord
