+++ src/useCases/GetPaymentNotificationMetadataUseCase.ts
@@ -1,74 +1,31 @@
-import * as E from 'fp-ts/Either';
 import * as TE from 'fp-ts/TaskEither';
 import * as RA from 'fp-ts/ReadonlyArray';
-import * as O from 'fp-ts/Option';
-import { flow, pipe } from 'fp-ts/function';
+import { pipe } from 'fp-ts/function';
 import {
   GetPaymentNotificationMetadataRecord,
-  makePaymentNotificationAttachmentDownloadMetadataResponse,
-} from '../domain/GetPaymentNotificationMetadataRecordRepository';
-import { ApiKey } from '../generated/definitions/ApiKey';
-import { Iun } from '../generated/definitions/Iun';
-import { authorizeApiKey } from '../domain/authorize';
+  makeGetPaymentNotificationMetadataRecord,
+} from '../domain/GetPaymentNotificationMetadataRecord';
 import { computeSnapshot } from '../domain/Snapshot';
-import { NotificationPaymentAttachment } from '../generated/definitions/NotificationPaymentAttachment';
-import { NotificationPaymentInfo } from '../generated/definitions/NotificationPaymentInfo';
 import { isNewNotificationRecord } from '../domain/NewNotificationRecord';
 import { isCheckNotificationStatusRecord } from '../domain/CheckNotificationStatusRecord';
 import { isConsumeEventStreamRecord } from '../domain/ConsumeEventStreamRecord';
 import { SystemEnv } from './SystemEnv';
 
-// FIXME: The attachmentName type should be an enum -> check type generated by openapi
-const getNotificationPaymentAttachment =
-  (attachmentName: string) =>
-  (payment: NotificationPaymentInfo): O.Option<NotificationPaymentAttachment> => {
-    switch (attachmentName) {
-      case 'F24_FLAT':
-        return O.fromNullable(payment.f24flatRate);
-      case 'F24_STANDARD':
-        return O.fromNullable(payment.f24standard);
-      case 'PAGOPA':
-        return O.fromNullable(payment.pagoPaForm);
-      default:
-        return O.none;
-    }
-  };
-
 export const GetPaymentNotificationMetadataUseCase =
   (env: SystemEnv) =>
-  (apiKey: ApiKey) =>
-  (iun: Iun) =>
-  (recipientId: number) =>
-  (attachmentName: string): TE.TaskEither<Error, GetPaymentNotificationMetadataRecord['output']> =>
+  (apiKey: GetPaymentNotificationMetadataRecord['input']['apiKey']) =>
+  (iun: GetPaymentNotificationMetadataRecord['input']['iun']) =>
+  (recipientId: GetPaymentNotificationMetadataRecord['input']['recipientId']) =>
+  (
+    attachmentName: GetPaymentNotificationMetadataRecord['input']['attachmentName']
+  ): TE.TaskEither<Error, GetPaymentNotificationMetadataRecord['output']> =>
     pipe(
-      authorizeApiKey(apiKey),
-      E.map(() =>
-        pipe(
-          TE.of(computeSnapshot(env)),
-          TE.ap(pipe(env.recordRepository.list(), TE.map(RA.filterMap(isNewNotificationRecord)))),
-          TE.ap(pipe(env.recordRepository.list(), TE.map(RA.filterMap(isCheckNotificationStatusRecord)))),
-          TE.ap(pipe(env.recordRepository.list(), TE.map(RA.filterMap(isConsumeEventStreamRecord)))),
-          TE.map(
-            flow(
-              RA.filterMap(O.fromEither),
-              RA.chain((notification) => (notification.iun === iun ? notification.recipients : RA.empty)),
-              RA.filterMap((recipient) => O.fromNullable(recipient.payment)),
-              RA.findLastMap(getNotificationPaymentAttachment(attachmentName)),
-              O.map(makePaymentNotificationAttachmentDownloadMetadataResponse(env)),
-              O.map((paymentAttachment) => ({ statusCode: 200 as const, returned: paymentAttachment })),
-              O.getOrElseW(() => ({ statusCode: 404 as const, returned: undefined }))
-            )
-          )
-        )
-      ),
-      flow(E.sequence(TE.ApplicativePar), TE.map(E.toUnion)),
-      TE.map((output) => ({
-        type: 'GetPaymentNotificationMetadataRecord' as const,
-        input: { apiKey, iun, recipientId, attachmentName },
-        output,
-        loggedAt: env.dateGenerator(),
-      })),
-      TE.chain(env.getPaymentNotificationMetadataRecordRepository.insert),
+      TE.of(computeSnapshot(env)),
+      TE.ap(pipe(env.recordRepository.list(), TE.map(RA.filterMap(isNewNotificationRecord)))),
+      TE.ap(pipe(env.recordRepository.list(), TE.map(RA.filterMap(isCheckNotificationStatusRecord)))),
+      TE.ap(pipe(env.recordRepository.list(), TE.map(RA.filterMap(isConsumeEventStreamRecord)))),
+      TE.map(makeGetPaymentNotificationMetadataRecord(env)({ apiKey, iun, recipientId, attachmentName })),
+      TE.chain(env.recordRepository.insert),
       TE.map((record) => record.output)
     );
